+++ plugins/flexdmd/resources/Bitmap.h
@@ -12,13 +12,13 @@ public:
    Bitmap(void* pData, AssetType assetType);
    Bitmap(Bitmap* pCachedBitmap);
    ~Bitmap();
-   
+
    PSC_IMPLEMENT_REFCOUNT()
 
    int GetWidth() const;
    int GetHeight() const;
-   float GetLength() const;
-   float GetFrameDelay(int pos) const;
+   int GetLength() const;
+   int GetFrameDelay(int pos) const;
    int GetFrameCount() const;
    SDL_Surface* GetSurface() const;
    SDL_Surface* GetFrameSurface(int pos) const;
